Make tests for UserController
