+++ config/packages/test/dama_doctrine_test_bundle.yaml
@@ -1,0 +1,4 @@
+dama_doctrine_test:
+    enable_static_connection: true
+    enable_static_meta_data_cache: true
+    enable_static_query_cache: true
