+++ phpunit.xml.dist
@@ -39,11 +39,19 @@
             <php outputFile="coverage.php"/>
             <text outputFile="coverage.txt" showUncoveredFiles="true" showOnlySummary="true"/>
         </report>
+        <exclude>
+            <file>src/Kernel.php</file>
+            <directory>src/DataFixtures</directory>
+        </exclude>
     </coverage>
 
     <listeners>
         <listener class="Symfony\Bridge\PhpUnit\SymfonyTestsListener" />
     </listeners>
+
+    <extensions>
+        <extension class="DAMA\DoctrineTestBundle\PHPUnit\PHPUnitExtension" />
+    </extensions>
 
     <!-- Run `composer require symfony/panther` before enabling this extension -->
     <!--
