+++ tests/Controller/SecurityControllerTest.php
@@ -2,15 +2,31 @@
 
 namespace Tests\Controller;
 
+use App\Repository\UserRepository;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class SecurityControllerTest extends WebTestCase
 {
+    private KernelBrowser $client;
+
+    public function setUp(): void
+    {
+        $this->client = static::createClient();
+    }
+
+    public function userLogin()
+    {
+        $userRepository = static::getContainer()->get(UserRepository::class);
+
+        $testUser = $userRepository->findOneBy(["email" => 'user@email.com']);
+
+        $this->client->loginUser($testUser);
+    }
+
     public function testLogin(): void
     {
-        $client = static::createClient();
-
-        $client->request('GET', '/login');
+        $this->client->request('GET', '/login');
         
         $this->assertResponseStatusCodeSame(200);
         $this->assertSelectorTextContains('h1', 'Connectez-vous');
@@ -19,38 +35,48 @@ class SecurityControllerTest extends WebTestCase
 
     public function testTryToLoginWithInvalidCredentials()
     {
-        $client = static::createClient();
-        $crawler = $client->request('GET', '/login');
+        $crawler = $this->client->request('GET', '/login');
 
         $form = $crawler->selectButton('Se connecter')->form([
             '_username' => 'username',
             '_password' => 'badPassword'
         ]);
 
-        $client->submit($form);
+        $this->client->submit($form);
         $this->assertResponseRedirects('http://localhost/login');
 
-        $client->followRedirect();
+        $this->client->followRedirect();
 
         $this->assertSelectorExists('.alert.alert-danger');
     }
 
     public function testTryToLoginWithValidCredentials()
     {
-        $client = static::createClient();
-        $crawler = $client->request('GET', '/login');
+        $crawler = $this->client->request('GET', '/login');
 
         $form = $crawler->selectButton('Se connecter')->form([
             '_username' => 'simple_user',
             '_password' => 'secret'
         ]);
 
-        $client->submit($form);
+        $this->client->submit($form);
         $this->assertResponseRedirects('http://localhost/');
 
-        $client->followRedirect();
+        $this->client->followRedirect();
 
         $this->assertResponseStatusCodeSame(200);
         $this->assertSelectorTextContains('h1', "Bienvenue sur Todo List, l'application vous permettant de gérer l'ensemble de vos tâches sans effort !");
+    }
+
+    public function testLogout()
+    {
+        $this->userLogin();
+
+        $this->client->request('GET', '/logout');
+        $this->client->followRedirect();
+        $this->client->followRedirect();
+
+        $this->assertResponseStatusCodeSame(200);
+        $this->assertSelectorTextContains('h1', 'Connectez-vous');
     }
 }
