+++ tests/Controller/TaskControllerTest.php
@@ -2,15 +2,47 @@
 
 namespace Tests\Controller;
 
+use App\Entity\Task;
+use App\Entity\User;
+use App\Repository\UserRepository;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class TaskControllerTest extends WebTestCase
 {
+    private KernelBrowser $client;
+
+    private $entityManager;
+
+    public function setUp(): void
+    {
+        $this->client = static::createClient();
+
+        $this->entityManager = $this->client->getContainer()
+        ->get('doctrine')
+        ->getManager();
+    }
+
+    public function userLogin()
+    {
+        $userRepository = static::getContainer()->get(UserRepository::class);
+
+        $testUser = $userRepository->findOneBy(["email" => 'user@email.com']);
+
+        $this->client->loginUser($testUser);
+    }
+
+    public function adminLogin()
+    {
+        $userRepository = static::getContainer()->get(UserRepository::class);
+
+        $testUser = $userRepository->findOneBy(["email" => 'admin@email.com']);
+
+        $this->client->loginUser($testUser);
+    }
     public function testListTasksWithoutLogin(): void
     {
-        $client = static::createClient();
-
-        $client->request('GET', '/tasks');
+        $this->client->request('GET', '/tasks');
         
         $this->assertResponseStatusCodeSame(302);
         $this->assertResponseRedirects('http://localhost/login');
@@ -18,9 +50,7 @@ class TaskControllerTest extends WebTestCase
 
     public function testCreateTaskWithoutLogin(): void
     {
-        $client = static::createClient();
-
-        $client->request('GET', '/tasks/create');
+        $this->client->request('GET', '/tasks/create');
 
         $this->assertResponseStatusCodeSame(302);
         $this->assertResponseRedirects('http://localhost/login');
@@ -28,9 +58,7 @@ class TaskControllerTest extends WebTestCase
 
     public function testEditTaskWithoutLogin(): void
     {
-        $client = static::createClient();
-
-        $client->request('GET', '/tasks/1/edit');
+        $this->client->request('GET', '/tasks/1/edit');
 
         $this->assertResponseStatusCodeSame(302);
         $this->assertResponseRedirects('http://localhost/login');
@@ -38,9 +66,7 @@ class TaskControllerTest extends WebTestCase
 
     public function testToogleTaskWithoutLogin(): void
     {
-        $client = static::createClient();
-
-        $client->request('GET', '/tasks/1/toggle');
+        $this->client->request('GET', '/tasks/1/toggle');
 
         $this->assertResponseStatusCodeSame(302);
         $this->assertResponseRedirects('http://localhost/login');
@@ -48,11 +74,142 @@ class TaskControllerTest extends WebTestCase
 
     public function testDeleteTaskWithoutLogin(): void
     {
-        $client = static::createClient();
-
-        $client->request('GET', '/tasks/1/delete');
+        $this->client->request('GET', '/tasks/1/delete');
 
         $this->assertResponseStatusCodeSame(302);
         $this->assertResponseRedirects('http://localhost/login');
+    }
+
+    public function testCreateTask()
+    {
+        $this->userLogin();
+
+        $crawler = $this->client->request('GET', '/tasks/create');
+
+        $form = $crawler->selectButton('Ajouter')->form([
+            'task' => [
+                'title' => 'Une nouvelle tâche',
+                'content' => 'pour tester testCreateTask()'
+            ]
+        ]);
+
+        $this->client->submit($form);
+        $this->assertResponseRedirects('/tasks');
+
+        $this->client->followRedirect();
+
+        $this->assertResponseStatusCodeSame(200);
+        $this->assertSelectorExists('.alert.alert-success');
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+    }
+
+    public function testEditTask()
+    {
+        $this->userLogin();
+
+        $crawler = $this->client->request('GET', '/');
+
+        $crawler = $this->client->clickLink('Consulter la liste des tâches à faire');
+
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+
+        $crawler = $this->client->clickLink("La tâche d'un utilisateur classique");
+
+        $this->assertSelectorTextContains('h1', "Modifier La tâche d'un utilisateur classique");
+
+        $form = $crawler->selectButton('Modifier')->form([
+            'task[title]' => "La tâche d'un utilisateur classique",
+            'task[content]' => "Elle n'est pas si difficile"
+        ]);
+
+        $this->client->submit($form);
+        $this->client->followRedirect();
+
+        $this->assertResponseStatusCodeSame(200);
+        $this->assertSelectorExists('.alert.alert-success');
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+    }
+
+    public function testToggleTask()
+    {
+        $this->userLogin();
+
+        $this->client->request('GET', '/');
+
+        $crawler = $this->client->clickLink('Consulter la liste des tâches à faire');
+
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+
+        $form = $crawler->filter('form')->selectButton('Marquer comme faite')->eq(5)->form();
+        $crawler = $this->client->submit($form);
+
+        $this->client->followRedirect();
+        $this->assertSelectorExists('div.alert.alert-success');
+    }
+
+    public function testDeleteTask()
+    {
+        $this->userLogin();
+
+        $this->client->request('GET', '/');
+
+        $crawler = $this->client->clickLink('Consulter la liste des tâches à faire');
+
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+
+        $form = $crawler->filter('form')->selectButton('Supprimer')->eq(4)->form();
+        $crawler = $this->client->submit($form);
+
+        $this->client->followRedirect();
+        $this->assertSelectorExists('div.alert.alert-success');
+    }
+
+    public function testDeleteTaskByAnotherUser()
+    {
+        $this->userLogin();
+
+        $this->client->request('GET', '/');
+
+        $crawler = $this->client->clickLink('Consulter la liste des tâches à faire');
+
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+
+        $form = $crawler->filter('form')->selectButton('Supprimer')->eq(2)->form();
+        $crawler = $this->client->submit($form);
+
+        $this->assertResponseStatusCodeSame(403);
+    }
+
+    public function testDeleteTaskOfAnonymousUserWithSimpleUserLogin()
+    {
+        $this->userLogin();
+
+        $this->client->request('GET', '/');
+
+        $crawler = $this->client->clickLink('Consulter la liste des tâches à faire');
+
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+
+        $form = $crawler->filter('form')->selectButton('Supprimer')->eq(1)->form();
+        $crawler = $this->client->submit($form);
+
+        $this->assertResponseStatusCodeSame(403);
+    }
+
+    public function testDeleteTaskOfAnonymousUserWithAdminLogin()
+    {
+        $this->adminLogin();
+
+        $this->client->request('GET', '/');
+
+        $crawler = $this->client->clickLink('Consulter la liste des tâches à faire');
+
+        $this->assertSelectorTextContains('h1', 'Liste des tâches');
+
+        $form = $crawler->filter('form')->selectButton('Supprimer')->eq(2)->form();
+        $crawler = $this->client->submit($form);
+
+        $this->client->followRedirect();
+        $this->assertSelectorExists('div.alert.alert-success');
     }
 }
